feat: create auth mapper

+++ src/main/java/com/osporo/engine/auth/dto/LoginResponse.java
@@ -10,6 +10,6 @@ public record LoginResponse(
         String refreshToken,
 
         @JsonProperty("expires_in")
-        int expiresIn
+        Long expiresIn
 ) {
 }
